🔨 [functional] Move `install` tests to separate module
